cli/new: inline prompt factories and move to param collection to preparation
